Update AutoPackOpener.ahk
Removed WinActive to focus the game window, which was causing lag and being disruptive and overall breaking the script timings. 

Add back in the coordinates for the tooltip to be pinned to the bottom of the screen out of the way of the cards.

Tested now for three full 4 hand 8 pack draws worked all times without issue now.